Add Bluetooth functionality and command processing for remote control
- Implement BluetoothConnection interface and its BLE variant.
- Create CommandProcessor for handling commands related to thermostat, humidifier, lighting, and turnstile.
- Add RemoteControlGateway to manage Bluetooth communication and state publishing.
- Introduce unit tests for Bluetooth command processing and remote gateway functionality.
- Update Arduino mock library to support String operations and serial input handling.
- Modify run_tests.sh for compatibility with Arduino UNO R4 WiFi.
- Update Wokwi configuration for new build paths.
- Document Bluetooth module and command protocol in README.

+++ lib/humidity/humidifier.h
@@ -7,12 +7,12 @@
 
 class HumidifierControl {
     private:
-        int _humidifierPin;
+        uint8_t _humidifierPin;
         float _targetHumidity;
 
     public:
         HumidifierControl(int humidifierPin, float targetHumidity)
-            : _humidifierPin(humidifierPin), _targetHumidity(targetHumidity) {}
+            : _humidifierPin(static_cast<uint8_t>(humidifierPin)), _targetHumidity(targetHumidity) {}
 
         void begin() {
             beginHumiditySensor(); // Initialize the DHT sensor
